Apply suggestion from @ribbanya

+++ src/melee/it/items/itmatadogas.c
@@ -68,7 +68,7 @@ void it_802CB2B0(Item_GObj* gobj)
     if (ip->xDAC_itcmd_var0 != 0) {
         itMatadogasAttributes* attrs = ip->xC4_article_data->x4_specialAttributes;
         if (--ip->xDD4_itemVar.matadogas.x60 == 0) {
-            ip->xDD4_itemVar.matadogas.x60 = (s32) attrs->x4;
+            ip->xDD4_itemVar.matadogas.x60 = attrs->x4;
             if (HSD_Randi(2) == 0) {
                 it_802CB4F0(gobj, 0xC1, attrs->x8);
             } else {
